Fix compatibility with React 16.4
React 16.4 included a change to the behavior of `getDerivedStateFromProps` which is now called after internal state updates in addition to received props. This change updates the comparison logic between current and incoming date values to ensure the local state is not overwritten with the prop value immediately following a state update.